fix: change studentId type into long

+++ src/main/java/net/bnijik/schooldbcli/dao/course/CourseDao.java
@@ -7,5 +7,5 @@ import net.bnijik.schooldbcli.model.Course;
 import java.util.List;
 
 public interface CourseDao extends Dao<Course> {
-    List<Course> findAllForStudent(int studentId, Page page);
+    List<Course> findAllForStudent(long studentId, Page page);
 }
